requirements.txt
pyspark==3.5.1          # Apache Spark for big data paraller processing 
apache-airflow-providers-apache-spark   # Airflow provider for Spark  
pandas                  # Data manipulation and analysis  
polars                  # Fast DataFrame library
adbc-driver-postgresql  # Polars engine
sqlalchemy              # SQL toolkit and Object Relational Mapper
psycopg2-binary         # PostgreSQL support
pymysql                 # MySQL support
pyarrow                 # Parquet/Arrow support
fastparquet             # Parquet support
openpyxl                # Excel file support
requests                # HTTP requests
boto3                   # AWS SDK
google-cloud-storage    # Google Cloud Storage SDK
azure-storage-blob      # Azure Blob Storage SDK
pyyaml                  # YAML config support
python-dotenv           # .env file support
loguru                  # Logging
jupyter                 # Jupyter notebooks
plotly                  # Visualization
pydantic                # Data validation and settings management
kafka-python            # Apache Kafka client
beautifulsoup4          # HTML/XML parsing
flake8                  # Linting for style and errors
black                   # Code formatter
isort                   # Import sorting
sqlfluff                # SQL linter and formatter
pre-commit              # Framework for pre-commit hooks
python-dotenv           # For loading environment